+++ test/integration/run.sh
@@ -27,11 +27,12 @@ export GLACIER_TESTS=${GLACIER_TESTS:-0}
 
 export AZURE_TESTS=${AZURE_TESTS:-1}
 export RUN_ADVANCED_TESTS=${RUN_ADVANCED_TESTS:-1}
+export GODEBUG=${GODEBUG:-}
 export S3_DEBUG=${S3_DEBUG:-false}
 export GCS_DEBUG=${GCS_DEBUG:-false}
 export FTP_DEBUG=${FTP_DEBUG:-false}
 export SFTP_DEBUG=${SFTP_DEBUG:-false}
-export GODEBUG=${GODEBUG:-}
+export AZBLOB_DEBUG=${AZBLOB_DEBUG:-false}
 export CLICKHOUSE_DEBUG=${CLICKHOUSE_DEBUG:-false}
 
 if [[ "${CLICKHOUSE_VERSION}" == 2* || "${CLICKHOUSE_VERSION}" == "head" ]]; then